Inventory Overhaul part 1
Fixed a bug in loading correct preset inside the main DND_Hash function.

Reorganised the command registry.

Remove the turn, repeatTurn, and initiative commands and systems. gearing up to replace the combat entity system with narrative injury based system.

Completely overhauled doTake, doDrop, and doReward to work with story cards and their item format.

doEquip semi-working working with new item format.

Changed the pluralize function to singularize so that the boolean argument reads more clearly when calling.

Added a testing command that runs test cases for functions. Currently only for doTake and doDrop; but can be expanded to any input function.

Add a function to compare words with plural and in lower case. (easier to call for any case we wish to do so)

Added a function to build story cards. Credits to lewd leah, and her auto cards project, from which this code was sourced.

Fixed output.js; re-added cases that had been removed accidently.