Reworked doBuy
Unified parseQuantityAndName to work for take, drop, give, and buy. (updated those functions)

Reworked the doBuy function; with test cases.

Documented the getArguments, getArgument, and getArgumentRemainder functions. Also added an optional with_command argument to getArguments; so that it can be called with or without the command_name being returned as the first element. (true by default)